Add release checklist
This is a (hopefully) complete guide to doing an FCOS release as of
today. Lots of warts and rough bits. Though I'm interested in kind of
using it too as a living document to track the next steps in
streamlining the process (hence the couple of "in the future" items in
there).