updates: rollout next on 06/12 and testing on 06/15
Since `next` is a little more risky of a stream we can go ahead and
roll it out on Friday. For `testing` we'll wait til Monday. Since
the content in both releases is identical rolling out `next` early
will give us an early indication of if there is a problem.