rollout.py: check whether rollout matches latest release
This is an easy way to sanity-check that the build IDs match up. The
version we rollout should almost always match what the latest release on
that stream is. Add a check for this.

So... this script now does networking. Which should be fine in most
cases, but add a switch in case a user doesn't want to depend on the
network.