templates: tweak example `./rollout.py` invocation
It was 9 AM EDT, and I ran `./rollout.py add next ... "10 am ET"` and
for some reason it set the rollout to 10 AM of the following day instead
of the same day. Adding "today" fixed that.

Since in the majority of cases, the releaser will want the rollout to
start the same day, let's just incorporate that in the example
invocation.

This is possibly a bug with `'PREFER_DATES_FROM': 'future'`.