README.md: add some explanatory text with links, move release details
Let's have the README.md provide a short overview of what this repo is
about and move all the bits about actually cutting a release to a
separate dedicated document.

Mention update metadata since it's not just stream metadata.

Put less emphasis on the generator since it's an implementation detail
of cutting releases.