RELEASE: create new "standby executor" role
With a large number of release executors, it's difficult for any one of
them to build up a lot of experience with the release process.  There's
also no well-defined plan for assigning an executor to ad-hoc releases;
we've been doing it ad hoc.  Address both issues by creating a "standby
release executor" role to handle ad-hoc releases, and moving some of the
more experienced executors over to it.