workflows: add job to check stream metadata for unexpected changes
Stream metadata updates are large diffs containing mostly routine changes,
with a small chance of serious breakage.  Reviewing them is tedious and
reviewers have overlooked problems in the past.  To aid PR reviews, add a
CI job that "genericizes" stream metadata from both the base branch and
the PR (by replacing strings that routinely change) and then compares the
two.  If there are any remaining differences, fail the job (which will
not block merge) and add annotation boxes to the GitHub diff view that
mark the unexpected differences.