workflows: don't warn on expected fedora-coreos-stream-generator bumps
We currently warn on fedora-coreos-stream-generator version changes,
even routine ones, and the false positive can lead to confusion.
Assume that stream metadata is built with the latest tag of
fedora-coreos-stream-generator.  For all streams that have been changed
in the current PR, update the old genericized metadata to reference the
new tag, avoiding the change warning.  For streams that haven't, we can
skip the comparison entirely.

This will still warn if stream metadata is built from an old generator
tag, or from a generator commit that has not been tagged.