workflows/rollout: don't bump metadata timestamps of unchanged streams
After regenerating stream metadata for an unchanged stream, we skip
committing it, but leave the regenerated metadata in the work tree with
an updated timestamp.  The create-pull-request action notices those
changes and commits them with a generic commit title
("[create-pull-request] automated change").

Discard any uncommitted changes to avoid the spurious automatic commit.

Fixes: 84b8393c86 ("workflows/rollout: switch to create-pull-request action")