updates/next: switch reason for 42.20250914.1.0 barrier
It just so happens that there are two reasons to make 42.20250914.1.0
a barrier. It's the last `next` release of F42 AND it includes the
code for migrating a system to container signature verification. Let's
include the link to the signature verification migration here since
that one is probably a little more significant.